Fixed all comments and added all mobile footer css on footer file
